save limit 100mb, fix save workspace rough, add unit tests, fix user tests, add shared worspace types, new component for user settings, add new workspace logic, add local cache for reloading page

+++ backend/tests/setup.ts
@@ -10,7 +10,8 @@ beforeAll(async () => {
 
 afterEach(async () => {
   // wichtig: zwischen jedem Test die Collections leeren
-  const collections = await mongoose.connection.db!.collections();
+  if (!mongoose.connection.db) return;
+  const collections = await mongoose.connection.db.collections();
   for (const c of collections) {
     await c.deleteMany({});
   }
